# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: "3API Relay Panel"
  version: "0.5.0"
  description: "Anthropic-compatible LLM relay with multi-tenant storefront, reseller admin, plans / orders / subscriptions, webhook delivery, and per-tenant BYOK upstream channels."
servers:
  - url: "https://3api.pro"
    description: "SaaS root (multi-tenant)"
  - url: "https://{store}.3api.pro"
    description: "Tenant storefront (replace {store})"
  - url: "http://localhost:8080"
    description: "Local dev"
...
components:
  securitySchemes:
    apiKey:
      type: "http"
      scheme: "bearer"
      bearerFormat: "sk-..."
    adminBearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    customerBearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
paths:
...
  /admin/me:
    get:
      summary: "Admin + tenant + brand bundle"
      tags:
        - "admin"
      responses:
        "200":
          description: "Bundle"
      security:
        - adminBearer: []
...
  /admin/end-users:
    get:
      summary: "List end users (paginated)"
      tags:
        - "admin"
      responses:
        "200":
          description: "End users"
      security:
        - adminBearer: []
    post:
      summary: "Create an end user"
      tags:
        - "admin"
      responses:
        "201":
          description: "Created"
        "409":
          description: "email already exists"
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "bob@example.com"
              password: "secret"
              initial_quota_cents: 10000
  /admin/plans:
    get:
      summary: "List plans"
      tags:
        - "admin"
      responses:
        "200":
          description: "Plans"
      security:
        - adminBearer: []
    post:
      summary: "Create a plan"
      tags:
        - "admin"
      responses:
        "201":
          description: "Plan created"
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Pro"
              slug: "pro"
              period_days: 30
              quota_tokens: 10000000
              price_cents: 9900
...
  /admin/channels:
    get:
      summary: "List upstream channels"
      tags:
        - "admin"
      responses:
        "200":
          description: "Channels"
      security:
        - adminBearer: []
    post:
      summary: "Create upstream channel"
      tags:
        - "admin"
      responses:
        "201":
          description: "Created"
      security:
        - adminBearer: []
  /admin/wholesale:
    get:
      summary: "Wholesale balance"
      tags:
        - "admin"
      responses:
        "200":
          description: "Balance"
      security:
        - adminBearer: []
  /admin/affiliate:
    get:
      summary: "Affiliate stats"
      tags:
        - "admin"
      responses:
        "200":
          description: "Stats + invite link"
      security:
        - adminBearer: []
...
  /admin/system-setting:
    get:
      summary: "Read tenant runtime switches"
      tags:
        - "admin"
      responses:
        "200":
          description: "Settings"
      security:
        - adminBearer: []
    patch:
      summary: "Update tenant runtime switches"
      tags:
        - "admin"
      responses:
        "200":
          description: "Updated"
      security:
        - adminBearer: []
...
  /admin/webhooks:
    get:
      summary: "List webhook subscriptions"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "200":
          description: "Webhook list incl. last_triggered_at, fail_count_total"
      security:
        - adminBearer: []
    post:
      summary: "Create webhook subscription"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "201":
          description: "Created (secret returned once)"
      description: "Returns the HMAC secret in the response **once** — record it for signature verification. Subsequent reads will not include the secret. Allowed event types: order.paid, subscription.expired, refund.processed, wholesale.low."
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              url: "https://example.com/3api-webhook"
              events:
                - "order.paid"
                - "subscription.expired"
...